NMS-9771: Restructure and improve Help/Support pages
The support page contains the web form which is used by OpenNMS administrators to open commercial support tickets.
The account. This page is now restricted to Admin users. Introduced a Help page which gives information where to find
documentation and to get in contact with people. To give users a better visibility what version of OpenNMS is running,
the version number is added to the footer page. To make it easier getting version information about the running Open
"About" page in the user's menu is added. The web site components are refactored to be reusable as JSP includes.

NMS-9771: Fix review comments

* Fix single import in bootstrap-footer.jsp
* Fix test to download source code from GitHub
* Hyphenate "support-friendly"